+++ frontend/sign_in.html
@@ -9,8 +9,8 @@
         @import url('https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');
     </style>
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.7.2/css/all.min.css">
-    <script src='src/modules/signin.js'></script>
     
+
 </head>
 
 <body>
@@ -31,8 +31,11 @@
             <div>
                 <p>¿Ya tienes cuenta?</p>
             </div> 
-            <button class="button" id="signin">Inicia sesión</button>
+            <button id="log-in-button" class="button" >Inicia sesión</button>
         </div>
     </div>
+    
+    <script type="module" src='./src/signin.js' ></script>
 </body>
+
 </html>
