+++ frontend/src/login.js
@@ -2,6 +2,7 @@ import * as backend from "./modules/backend_connection.js"
 import { requestFeedback } from "./modules/ui_feedback.js"
 import { redirectTo } from "./modules/user_fetch.js"
 
+
 //DOM Ids
 const IDs = {
     loginUsername: "login-user-name",
